+++ __tests__/ui/roles/RolesAdd.unit.test.js
@@ -1,5 +1,5 @@
 import { useStrict } from 'mobx'
-import RolesAdd from '~/ui/role/RolesAdd'
+import RolesAdd from '~/ui/roles/RolesAdd'
 
 let props
 let wrapper
