+++ app/javascript/ui/organizations/OrganizationAvatar.js
