+++ app/javascript/ui/pages/CollectionPage.js
@@ -10,8 +10,8 @@ import Header from '~/ui/layout/Header'
 import PageContainer from '~/ui/layout/PageContainer'
 import CollectionGrid from '~/ui/grid/CollectionGrid'
 import Breadcrumb from '~/ui/layout/Breadcrumb'
-import RolesSummary from '~/ui/role/RolesSummary'
-import RolesMenu from '~/ui/role/RolesMenu'
+import RolesSummary from '~/ui/roles/RolesSummary'
+import RolesMenu from '~/ui/roles/RolesMenu'
 import EditableName from './shared/EditableName'
 
 const isHomepage = ({ path }) => path === '/'
