+++ app/javascript/ui/roles/RoleSelect.js
@@ -11,7 +11,7 @@ import {
   Row,
   RowItemLeft,
 } from '~/ui/global/styled/layout'
-import UserAvatar from '~/ui/user/UserAvatar'
+import UserAvatar from '~/ui/users/UserAvatar'
 
 const materialStyles = {
   root: {
