+++ app/javascript/ui/users/UserAvatar.js
